refactor: Move getLocalFilename() into new MediaItem class
